refactor: extract shared JSON output code to json_output.hpp
