+++ barretenberg/cpp/src/barretenberg/api/api_ultra_honk.cpp
@@ -1,96 +1,25 @@
 #include "api_ultra_honk.hpp"
 
 #include "barretenberg/api/file_io.hpp"
+#include "barretenberg/api/json_output.hpp"
 #include "barretenberg/bbapi/bbapi_ultra_honk.hpp"
 #include "barretenberg/common/bb_bench.hpp"
 #include "barretenberg/common/get_bytecode.hpp"
 #include "barretenberg/common/map.hpp"
 #include "barretenberg/common/throw_or_abort.hpp"
-#include "barretenberg/common/version.hpp"
 #include "barretenberg/dsl/acir_format/acir_to_constraint_buf.hpp"
 #include "barretenberg/dsl/acir_proofs/honk_contract.hpp"
 #include "barretenberg/dsl/acir_proofs/honk_optimized_contract.hpp"
 #include "barretenberg/dsl/acir_proofs/honk_zk_contract.hpp"
 #include "barretenberg/honk/proof_system/types/proof.hpp"
 #include "barretenberg/numeric/uint256/uint256.hpp"
-#include "barretenberg/serialize/msgpack_impl.hpp"
 #include "barretenberg/special_public_inputs/special_public_inputs.hpp"
 #include "barretenberg/srs/global_crs.hpp"
-#include <iomanip>
 #include <optional>
-#include <sstream>
 
 namespace bb {
 
 namespace {
-
-/**
- * @brief Convert a vector of uint256_t to a vector of hex strings
- */
-inline std::vector<std::string> uint256_vector_to_hex_strings(const std::vector<uint256_t>& fields)
-{
-    std::vector<std::string> result;
-    result.reserve(fields.size());
-    for (const auto& field : fields) {
-        std::stringstream ss;
-        ss << field; // uint256_t's operator<< outputs "0x" prefix
-        result.push_back(ss.str());
-    }
-    return result;
-}
-
-/**
- * @brief Convert bytes to a hex string with 0x prefix
- */
-inline std::string bytes_to_hex_string(const std::vector<uint8_t>& bytes)
-{
-    std::stringstream ss;
-    ss << "0x" << std::hex << std::setfill('0');
-    for (const auto& byte : bytes) {
-        ss << std::setw(2) << static_cast<int>(byte);
-    }
-    return ss.str();
-}
-
-/**
- * @brief Serializable structure for JSON output (msgpack-compatible)
- */
-struct JsonOutput {
-    std::vector<std::string> fields;
-    std::string vk_hash;   // Only for VK and proof (hash of VK the proof targets)
-    std::string file_kind; // "vk", "proof", or "public_inputs"
-    std::string bb_version;
-    std::string scheme;
-    std::string verifier_target; // Optional
-
-    MSGPACK_FIELDS(fields, vk_hash, file_kind, bb_version, scheme, verifier_target);
-};
-
-/**
- * @brief Build JSON output string using msgpack serialization
- */
-std::string build_json_output(const std::vector<uint256_t>& fields,
-                              const std::string& file_kind,
-                              const API::Flags& flags,
-                              const std::string& hash = "")
-{
-    JsonOutput output{
-        .fields = uint256_vector_to_hex_strings(fields),
-        .vk_hash = hash,
-        .file_kind = file_kind,
-        .bb_version = BB_VERSION,
-        .scheme = flags.scheme,
-        .verifier_target = flags.verifier_target,
-    };
-
-    msgpack::sbuffer buffer;
-    msgpack::pack(buffer, output);
-    msgpack::object_handle oh = msgpack::unpack(buffer.data(), buffer.size());
-
-    std::stringstream ss;
-    ss << oh.get();
-    return ss.str();
-}
 
 void write_vk_outputs(const bbapi::CircuitComputeVk::Response& vk_response,
                       const std::filesystem::path& output_dir,
